refactor: decouple more tools and agent files from vscode
- DiagnosticsTool.ts: remove unused vscode import
- latexPreview.ts: duck-type vscode.Uri to { fsPath: string }, add
  injectable onError callback for error reporting
- toolEditApproval.ts: narrow PendingApprovalEntry back to vscode.Uri,
  wire onError callback
- LspTools.ts: move navigateToFirstError, fetchDiagnosticsForFile, and
  executeGlobalCommand into VscodeIntegration.ts; remove vscode import
- ExecutionsTool.ts: fix isDirectory(stats.type) call from prior refactor
- remoteAgentUtils.ts: move from src/agent/remote/ to src/frontend/agents/
  since it's purely VS Code UI code
